Raghav/mongo persistence logic (#8)
* Updated POJO for Mongo DB params and also included Spring Validator framework annotations

* Logic for object comparison for stock feeds for locations/stock values and persist collated information to Mongo collection

* Updated with method comments and added ReadMe file

* Updated POJO elements to be non-primitive

+++ stock-message-generator/src/main/java/com/sapient/retail/service/streamkafka/repository/StockRepository.java
@@ -7,6 +7,6 @@ import com.sapient.retail.service.streamkafka.model.ProductStockInfo;
  * ProductStockInfo Repository for Reactive MongoDB interactions
  * @author ragarora
  */
-public interface StockRepository extends ReactiveMongoRepository<ProductStockInfo, Integer> {
+public interface StockRepository extends ReactiveMongoRepository<ProductStockInfo, Long> {
 	
 }
